+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/v4/src/webkit_webview_platform.dart
@@ -4,8 +4,8 @@
 
 import 'package:webview_flutter_platform_interface/v4/webview_flutter_platform_interface.dart';
 
-import 'webkit_navigation_delegate.dart';
 import 'webkit_webview_controller.dart';
+import 'webkit_webview_cookie_manager.dart';
 
 /// Implementation of [WebViewPlatform] using the WebKit API.
 class WebKitWebViewPlatform extends WebViewPlatform {
@@ -21,5 +21,19 @@ class WebKitWebViewPlatform extends WebViewPlatform {
     PlatformNavigationDelegateCreationParams params,
   ) {
     return WebKitNavigationDelegate(params);
+  }
+
+  @override
+  WebKitWebViewWidget createPlatformWebViewWidget(
+    PlatformWebViewWidgetCreationParams params,
+  ) {
+    return WebKitWebViewWidget(params);
+  }
+
+  @override
+  WebKitWebViewCookieManager createPlatformCookieManager(
+    PlatformWebViewCookieManagerCreationParams params,
+  ) {
+    return WebKitWebViewCookieManager(params);
   }
 }