+++ script/tool/lib/src/firebase_test_lab_command.dart
@@ -59,7 +59,7 @@ class FirebaseTestLabCommand extends PackageLoopingCommand {
         help:
             'Device model(s) to test. See https://cloud.google.com/sdk/gcloud/reference/firebase/test/android/run for more info');
     argParser.addOption('results-bucket',
-        defaultsTo: 'gs://flutter_firebase_testlab_staging');
+        defaultsTo: 'gs://flutter_cirrus_testlab');
     argParser.addOption(
       kEnableExperiment,
       defaultsTo: '',