+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+* Updates minimum Flutter version to 2.10.
 * Added setMaxVideoResolution method to support different video resolutions while picking a video stream from a source with different tracks
 * Fixes violations of new analysis option use_named_constants.