+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Updates minimum Flutter version to 2.10.
+* Fixes violations of new analysis option use_named_constants.
+
 ## 2.3.9
 
 * Updates ExoPlayer to 2.18.1.